requirements.txt
asyncpg
flask
jieba
llama_index
llama-index-question-gen-guidance
llama-index-question-gen-openai
llama-index-readers-pdf-marker
llama-index-readers-llama-parse
llama-index-readers-smart-pdf-loader
llama-index-indices-managed-postgresml
llama-index-graph-stores-neo4j
llama-index-storage-index-store-postgres
llama-index-storage-docstore-postgres
llama-index-storage-docstore-mongodb
llama-index-storage-index-store-mongodb
llama-index-vector-stores-postgres
llama-index-vector-stores-pinecone
llama-index-vector-stores-weaviate
llama-index-vector-stores-mongodb
llama-index-vector-stores-redis
llama-index-extractors-entity
llama-index-extractors-marvin
llama-index-embeddings-huggingface
llama-index-embeddings-ollama
llama-index-embeddings-instructor
llama_index_llms_ollama
llama-index-llms-openai
llama-index-llms-lmstudio
llama-index-tools-database
llmsherpa
llama-parse
lxml
marker-pdf
pandas
python-dotenv
pgsql
psycopg2-binary
spacy
streamlit
unstructured
bs4
markdown